+++ jaguar_jwt/CHANGELOG.md
@@ -6,6 +6,7 @@
 + Fixed validation to reject token when current time equals the Expiry time.
 + Added support for custom payload name.
 + Added more validation unit tests.
++ Fixed generation of JWT to use correct Base64url Encoding.
 
 ## 2.1.2